client/requirements.txt
typer>=0.9.0
rich>=13.0.0
httpx>=0.25.0
pyyaml>=6.0
